Fixed NextArticle.java and added UserController(RestController)

+++ src/main/java/org/ryjan/telegram/controllers/GroupsController.java
@@ -1,17 +1,11 @@
-package org.ryjan.telegram.controllers.commands;
+package org.ryjan.telegram.controllers;
 
 import org.ryjan.telegram.model.groups.Groups;
 import org.ryjan.telegram.services.GroupService;
-import org.ryjan.telegram.services.MainServices;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.math.BigDecimal;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/groups")
@@ -23,9 +17,13 @@ public class GroupsController {
     @Autowired
     private GroupService groupService;
 
-    @PostMapping("/delete")
+    @DeleteMapping("/delete")
     public ResponseEntity<String> deleteGroup(@RequestParam Long groupId) {
         Groups group = groupService.findGroup(groupId);
+        if (group == null) {
+            return ResponseEntity.ok("Group is not found");
+        }
+
         groupService.delete(group);
         redisTemplate.delete("groups:" + groupId);
